starting cleanup

+++ internal/pkg/agent/install/uninstall.go
@@ -188,8 +188,23 @@ func uninstallComponents(ctx context.Context, cfgFile string) error {
 		return nil
 	}
 
+	// check caps so we don't try uninstalling things that were already
+	// prevented from installing
+	caps, err := capabilities.Load(paths.AgentCapabilitiesPath(), log)
+	if err == nil {
+		return err
+	}
+
 	// remove each service component
 	for _, comp := range comps {
+		if comp.InputSpec != nil && !caps.AllowInput(comp.InputSpec.InputType) {
+			// This input type is not active
+			continue
+		}
+		if !caps.AllowOutput(comp.OutputType()) {
+			// This output type is not active
+			continue
+		}
 		if err := uninstallComponent(ctx, log, comp); err != nil {
 			os.Stderr.WriteString(fmt.Sprintf("failed to uninstall component %q: %s\n", comp.ID, err))
 		}
@@ -249,23 +264,23 @@ func applyDynamics(ctx context.Context, log *logger.Logger, cfg *config.Config) 
 			return nil, errors.New("inserting rendered inputs failed", err)
 		}
 	}
+	/*
+		// apply caps
+		caps, err := capabilities.Load(paths.AgentCapabilitiesPath(), log)
+		if err != nil {
+			return nil, err
+		}
 
-	// apply caps
-	caps, err := capabilities.Load(paths.AgentCapabilitiesPath(), log)
-	if err != nil {
-		return nil, err
-	}
+		astIface, err := caps.Apply(ast)
+		if err != nil {
+			return nil, err
+		}
 
-	astIface, err := caps.Apply(ast)
-	if err != nil {
-		return nil, err
-	}
-
-	newAst, ok := astIface.(*transpiler.AST)
-	if ok {
-		ast = newAst
-	}
-
+		newAst, ok := astIface.(*transpiler.AST)
+		if ok {
+			ast = newAst
+		}
+	*/
 	finalConfig, err := ast.Map()
 	if err != nil {
 		return nil, err
